extract VizLink class
